% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{passed to shiny::runApp (e.g., host, port, launch.browser)}
}
\description{
Run the Shiny application
}
